+++ src/device/tsp.script/Script/ScriptEntityBaseCollection.cs
@@ -585,7 +585,7 @@ public class ScriptEntityBaseCollection<TItem>( NodeEntityBase node ) : System.C
             if ( FirmwareScriptBase.ScriptNameExists( this.SavedScriptNames, scriptName ) )
                 isNull = session.DeleteSavedScript( scriptName );
             else
-                isNull = session.NillScript( scriptName );
+                isNull = FirmwareManager.NillScript( session, scriptName );
         }
         return isNull;
     }